+++ lib/enums.dart
@@ -1,3 +1,5 @@
 enum TextSize { normal, medium, label, title, subTitle, large, xl }
 
 enum TextColors { primary, secondary, accent, success, danger, info, warning }
+
+enum ButtonType { text, icon, iconText, textIcon }